+++ ads-in.js
@@ -1,6 +1,7 @@
 module.exports = function (RED) {
   'use strict'
-   var util = require('util')
+  var util = require('util')
+  var debug = require('debug')('node-red-contrib-ads:adsInNode')
 
   function adsInNode(config) {
     RED.nodes.createNode(this, config)
@@ -14,14 +15,18 @@ module.exports = function (RED) {
       node.timezone = config.timezone
       node.inValue = config.inValue||'payload'
       node.useInputMsg = config.useInputMsg||false
+      debug('config:',node)
 
       node.onAdsData = function (handle){
+        debug('onAdsData:','node.id',node.id,'node.symname',node.symname,'handle.value',handle.value)
         var msg = {}
         RED.util.setMessageProperty(msg, node.inValue, handle.value)
         node.send(msg)
+        debug('onAdsData:','node.id',node.id,'node.symname',node.symname,'msg',msg)
       }
 
       this.on("input", function(msg) {
+        debug('input:',msg)
         var outMsg = {}
         if (node.useInputMsg) {
           outMsg = msg
@@ -30,6 +35,7 @@ module.exports = function (RED) {
         node.adsDatasource.read(node,function (handle){
           RED.util.setMessageProperty(outMsg, node.inValue, handle.value)
           node.send(outMsg)
+          debug('input:','node.id',node.id,'node.symname',node.symname,'outMsg',outMsg)
         })
       })