+++ ads-out.js
@@ -1,8 +1,9 @@
 module.exports = function (RED) {
   'use strict'
   var util = require('util')
+  var debug = require('debug')('node-red-contrib-ads:adsOutNode')
 
-  function adsInNode(config) {
+  function adsOutNode(config) {
     RED.nodes.createNode(this, config)
     var node = this
     node.adsDatasource = RED.nodes.getNode(config.datasource)
@@ -12,8 +13,10 @@ module.exports = function (RED) {
       node.bytelength = config.varSize
       node.timezone = config.timezone
       node.outValue = config.outValue||'payload'
+      debug('config:',node)
 
       this.on("input", function(msg) {
+        debug('adsNotificationNode:','onAdsData:','node.id',node.id,'node.symname',node.symname,'msg',msg)
         var value = RED.util.getMessageProperty(msg,node.outValue)
         if (value !== undefined) {
           node.adsDatasource.write(node,value)
@@ -24,5 +27,5 @@ module.exports = function (RED) {
       })
     }
   }
-  RED.nodes.registerType('ADS Output', adsInNode)
+  RED.nodes.registerType('ADS Output', adsOutNode)
 }