+++ ads-symbols.js
@@ -2,6 +2,7 @@ module.exports = function (RED) {
   'use strict'
   var util = require('util')
   var adsHelpers = require('./ads-helpers')
+  var debug = require('debug')('node-red-contrib-ads:adsSymbolsNode')
 
   function adsSymbolsNode(config) {
     RED.nodes.createNode(this, config)
@@ -14,9 +15,11 @@ module.exports = function (RED) {
         adstype: config.varTyp,
         data: config.data
       }
+      debug('config:',node)
 
 
       node.onData = function (data){
+        debug('onData:','node.id',node.id,'node.type',node.type,'data',data)
         const msg = {
           payload: data
         }
@@ -25,6 +28,7 @@ module.exports = function (RED) {
 
       this.on("input", function(msg) {
         
+        debug('input:','node.id',node.id,'node.type',node.type,'data',msg)
         var call = node.adsDatasource.getDatatyps
         if (node.adscfg.data == 'SYMBOLES') {
           call = node.adsDatasource.getSymbols
@@ -61,6 +65,7 @@ module.exports = function (RED) {
       })
 
       node.on('close', function () {
+        debug('close:','node.id',node.id,'node.type',node.type)
       })
     }
   }