+++ ads-system.js
@@ -1,6 +1,7 @@
 module.exports = function (RED) {
   'use strict'
   var util = require('util')
+  var debug = require('debug')('node-red-contrib-ads:adsSystemNode')
 
   function adsSystemNode(config) {
     RED.nodes.createNode(this, config)
@@ -12,6 +13,7 @@ module.exports = function (RED) {
         symname: config.varName,
         adstype: config.varTyp
       }
+      debug('config:',node)
 
 
       node.onData = function (data){
@@ -19,15 +21,18 @@ module.exports = function (RED) {
           payload: data
         }
         node.send(msg)
+        debug('onData:','node.id',node.id,'node.type',node.type,'msg',msg)
       }
 
       node.adsDatasource.systemRegister(node)
 
       this.on("input", function(msg) {
+        debug('input:','node.id',node.id,'node.type',node.type,'msg',msg)
         node.adsDatasource.systemUpdate(node)
       })
 
       node.on('close', function () {
+        debug('close:','node.id',node.id,'node.type',node.type,'enter')
         node.adsDatasource.systemUnregister(node)
       })
     }